Move UTXO finder to `MintingFlowRouter` component
Moves UTXO call that happens every 10 seconds to `MintingFlowRouter`. THanks to
that we can pass the utxos to the components that need them.
